refactor: 점수 계산 로직 분리
