+++ src/main/java/chess/model/board/ChessBoard.java
@@ -4,18 +4,12 @@ import chess.model.piece.Empty;
 import chess.model.piece.Piece;
 import chess.model.piece.Side;
 import chess.model.position.ChessPosition;
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 public class ChessBoard {
     private static final int KING_COUNT = 2;
-    private static final int PAWN_COUNT_WHEN_SOLID_IN_FILE = 1;
-    private static final double POINT_WHEN_SAME_FILE_PAWN = -0.5;
 
     private final Map<ChessPosition, Piece> board;
     private final Side turn;
@@ -58,10 +52,8 @@ public class ChessBoard {
         return findPiece(position).isEnemy(side);
     }
 
-    public Map<Side, Point> calculatePoints() {
-        return Arrays.stream(Side.values())
-                .filter(side -> !side.isEmpty())
-                .collect(Collectors.toMap(Function.identity(), this::calculatePoint));
+    public PointCalculator getPointCalculator() {
+        return new PointCalculator(Collections.unmodifiableMap(board));
     }
 
     private void validateTurn(final Piece sourcePiece) {
@@ -89,52 +81,11 @@ public class ChessBoard {
         board.put(sourcePosition, sourcePiece.catchTargetPiece(targetPiece));
     }
 
-    private boolean hasSameFilePawn(final ChessPosition position, final Set<ChessPosition> positions) {
-        return calculatePawnCountSameFile(position, positions) != PAWN_COUNT_WHEN_SOLID_IN_FILE;
-    }
-
     private int calculateKingCount() {
         return (int) board.values()
                 .stream()
                 .filter(Piece::isKing)
                 .count();
-    }
-
-    private long getSameFilePawnCount(final Side side) {
-        final Set<ChessPosition> positions = findAllSamSidePawns(side);
-
-        return positions.stream()
-                .filter(position -> hasSameFilePawn(position, positions))
-                .count();
-    }
-
-    private long calculatePawnCountSameFile(final ChessPosition position, final Set<ChessPosition> positions) {
-        return positions.stream()
-                .filter(position::hasSameFile)
-                .count();
-    }
-
-    private Point calculateTotalSum(final Side side) {
-        return board.values()
-                .stream()
-                .filter(piece -> piece.isSameSide(side))
-                .map(Piece::getPoint)
-                .reduce(Point.getDefaults(), Point::sum);
-    }
-
-    private Set<ChessPosition> findAllSamSidePawns(final Side side) {
-        return board.entrySet()
-                .stream()
-                .filter(entry -> entry.getValue().isSameSide(side))
-                .filter(entry -> entry.getValue().isPawn())
-                .map(Entry::getKey)
-                .collect(Collectors.toSet());
-    }
-
-    private Point calculatePoint(Side side) {
-        final Point totalSum = calculateTotalSum(side);
-        final long sameFilePawn = getSameFilePawnCount(side);
-        return totalSum.sum(Point.from(sameFilePawn * POINT_WHEN_SAME_FILE_PAWN));
     }
 
     private Piece findPiece(final ChessPosition position) {
